update run_scope()
Update the run_scope() R function that allows users to choose the location of SCOPE.m in their file directory when first run within an R session. The function saves the file path to an R environment variable for subsequent function calls and removes the variable when the session is terminated.